pyproject.toml
[tool.black]
line-length = 120
target-version = ['py310']
force-exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
  | venv
  | .venv
  | migrations
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
skip_glob = ['**/migrations/**', '**/.venv/**']


[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["bekzod-fayzikuloff <fayzikuloff.b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
Markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
django-jazzmin = "^2.5.0"
celery = "^5.2.7"
redis = "^4.3.4"
django-celery-results = "^2.4.0"
django-dirtyfields = "^1.8.2"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.1"}
drf-spectacular = {extras = ["sidecar"], version = "^0.24.2"}
drf-spectacular-sidecar = "^2022.10.1"
Pillow = "^9.3.0"
django-cors-headers = "^3.13.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
Faker = "^14.0.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
